Get data from the INE API

A short description...

Usage

get_ine_data(
  indicator,
  lang = "PT",
  expected.duration = FALSE,
  max_cells = 30000,
  ...
)

Arguments

indicator

INE indicator ID as a 7 character string. Example: "0010003".

lang

Only "PT" implemented.

expected.duration

If TRUE, prints the expected remaining duration in the console.

max_cells

Integer smaller than or equal to 40000, the maximum number of cells allowed in each API call. Default value of 30000

...

Values for each Dimension of the indicator. Each parameter should be in the form dimN, with N one of {1, ..., Total number of dimensions}. If one of the dimensions is not included, output includes all values

Value

The desired values

Details

Logs the expected remaining duration in the console. Can be very high for indicators with many dimensions and/or many unique values. (E.g. Indicator "0008206 - Deaths by Place of residence, Sex, Age group and Death cause, Annual" will take many hours to extract completely)

Examples

get_ine_data("0010003") # A simple example
#>    dim_1 geocod   geodsg valor
#> 1   2010     PT Portugal 58.31
#> 2   2011     PT Portugal 60.04
#> 3   2012     PT Portugal 60.53
#> 4   2013     PT Portugal 58.92
#> 5   2014     PT Portugal  62.1
#> 6   2015     PT Portugal 61.28
#> 7   2016     PT Portugal 64.61
#> 8   2017     PT Portugal 64.55
#> 9   2018     PT Portugal 67.41
#> 10  2019     PT Portugal 69.16
#> 11  2020     PT Portugal 58.01
#> 12  2021     PT Portugal 60.72
#> 13  2022     PT Portugal 64.18
#> 14  2023     PT Portugal 66.08
get_ine_data("0008206", dim1 = "S7A1996", dim2 = c("11","111"),
             dim4 = c(1,19), dim5 = "TLES") # A more complex example
#>    dim_1 geocod     geodsg dim_3 dim_3_t dim_4        dim_4_t dim_5
#> 1   1996     11      Norte     2       M     1 Menos de 1 ano  TLES
#> 2   1996     11      Norte     1       H     1 Menos de 1 ano  TLES
#> 3   1996     11      Norte     T      HM     1 Menos de 1 ano  TLES
#> 4   1996     11      Norte     2       M    19 85 e mais anos  TLES
#> 5   1996     11      Norte     T      HM    19 85 e mais anos  TLES
#> 6   1996     11      Norte     1       H    19 85 e mais anos  TLES
#> 7   1996    111 Alto Minho     1       H     1 Menos de 1 ano  TLES
#> 8   1996    111 Alto Minho     2       M     1 Menos de 1 ano  TLES
#> 9   1996    111 Alto Minho     T      HM     1 Menos de 1 ano  TLES
#> 10  1996    111 Alto Minho     T      HM    19 85 e mais anos  TLES
#> 11  1996    111 Alto Minho     2       M    19 85 e mais anos  TLES
#> 12  1996    111 Alto Minho     1       H    19 85 e mais anos  TLES
#>                     dim_5_t valor
#> 1  Todas as causas de morte   142
#> 2  Todas as causas de morte   196
#> 3  Todas as causas de morte   338
#> 4  Todas as causas de morte  4832
#> 5  Todas as causas de morte  7154
#> 6  Todas as causas de morte  2322
#> 7  Todas as causas de morte     7
#> 8  Todas as causas de morte     7
#> 9  Todas as causas de morte    14
#> 10 Todas as causas de morte   799
#> 11 Todas as causas de morte   537
#> 12 Todas as causas de morte   262