Get indicator metadata

get_metadata() calls the metadata API.

Be aware that this function will still return a value and not throw an error if the indicator does not exist.

Usage

get_metadata(indicator, lang = "PT")

Arguments

indicator

INE indicator ID as a 7 character string. Example: "0010003".

lang

Only "PT" implemented.

Value

API response body as list.

See also

Examples

get_metadata("0010003")
#> $IndicadorCod
#> [1] "0010003"
#> 
#> $IndicadorNome
#> [1] "Percentagem do orçamento de Estado financiado por impostos cobrados internamente (%); Anual - Direção Geral do Orçamento (Ministério das Finanças)"
#> 
#> $Periodic
#> [1] "Anual"
#> 
#> $PrimeiroPeriodo
#> [1] "2010"
#> 
#> $UltimoPeriodo
#> [1] "2023"
#> 
#> $UnidadeMedida
#> [1] "Percentagem (%)"
#> 
#> $Potencia10
#> [1] "0"
#> 
#> $PrecisaoDecimal
#> [1] "2"
#> 
#> $Lingua
#> [1] "PT"
#> 
#> $DataUltimaAtualizacao
#> [1] "2022-12-28"
#> 
#> $DataExtracao
#> [1] "2023-05-05T16:25:42.790+01:00"
#>