Get data from the INE API

Return tidy data frames from Instituto Nacional de Estatística API

Usage

get_ine_data(
  indicator,
  lang = "PT",
  expected.duration = FALSE,
  max_cells = 30000,
  ...
)

Arguments

indicator

INE indicator ID as a 7 character string. Example: "0010003".

lang

One of "PT" or "EN". Default is "PT".

expected.duration

If TRUE, prints the expected remaining duration in the console.

max_cells

Integer smaller than or equal to 40000, the maximum number of cells allowed in each API call. Default value of 30000

...

Values for each Dimension of the indicator. Each parameter should be in the form dimN, with N one of {1, ..., Nº of dimensions}. If one of the dimensions is not included, output includes all values for that dimension.

Value

tidy data frame for the given indicator.

Details

Extraction time can be very high (many hours) for indicators with many dimensions and/or many unique values. E.g. Indicator "0008206 - Deaths by Place of residence, Sex, Age group and Death cause, Annual" will take many hours to extract.

See also

Examples

#A simple example. Returns the complete dataset.
get_ine_data("0011823")
#>   dim_1 geocod   geodsg dim_3
#> 1  2022     PT Portugal     1
#> 2  2022     PT Portugal     2
#> 3  2022     PT Portugal     3
#> 4  2022     PT Portugal     4
#> 5  2022     PT Portugal     5
#>                                                                                                                dim_3_t
#> 1                                                      Websites ou aplicações pouco intuitivos ou difíceis de utilizar
#> 2                                                       Problemas técnicos no funcionamento dos websites ou aplicações
#> 3 Problemas com a assinatura eletrónica ou com a identificação necessária para efetuar a autenticação e usar o serviço
#> 4                                                                   Falta de acesso aos métodos de pagamento admitidos
#> 5                                                              Serviços não acessíveis através de smartphone ou tablet
#>   valor
#> 1  33.8
#> 2  36.6
#> 3  15.4
#> 4  10.3
#> 5  14.4
# A more complex example. Returns a subset of the dataset.
# \donttest{
get_ine_data("0008206", dim1 = "S7A1996", dim2 = c("11","111"),
             dim4 = c(1,19), dim5 = "TLES")
#>    dim_1 geocod     geodsg dim_3 dim_3_t dim_4        dim_4_t dim_5
#> 1   1996     11      Norte     2       M     1 Menos de 1 ano  TLES
#> 2   1996     11      Norte     T      HM     1 Menos de 1 ano  TLES
#> 3   1996     11      Norte     1       H     1 Menos de 1 ano  TLES
#> 4   1996     11      Norte     2       M    19 85 e mais anos  TLES
#> 5   1996     11      Norte     T      HM    19 85 e mais anos  TLES
#> 6   1996     11      Norte     1       H    19 85 e mais anos  TLES
#> 7   1996    111 Alto Minho     2       M     1 Menos de 1 ano  TLES
#> 8   1996    111 Alto Minho     T      HM     1 Menos de 1 ano  TLES
#> 9   1996    111 Alto Minho     1       H     1 Menos de 1 ano  TLES
#> 10  1996    111 Alto Minho     T      HM    19 85 e mais anos  TLES
#> 11  1996    111 Alto Minho     1       H    19 85 e mais anos  TLES
#> 12  1996    111 Alto Minho     2       M    19 85 e mais anos  TLES
#>                     dim_5_t valor
#> 1  Todas as causas de morte   142
#> 2  Todas as causas de morte   338
#> 3  Todas as causas de morte   196
#> 4  Todas as causas de morte  4832
#> 5  Todas as causas de morte  7154
#> 6  Todas as causas de morte  2322
#> 7  Todas as causas de morte     7
#> 8  Todas as causas de morte    14
#> 9  Todas as causas de morte     7
#> 10 Todas as causas de morte   799
#> 11 Todas as causas de morte   262
#> 12 Todas as causas de morte   537
# }